# misc small perf improvments