```
                                150982683ad1dcc208:  ecccafff989f6d06f7:
           django_template3.py             2.8s (4)             2.8s (4)  -0.9%
                 pyxl_bench.py             2.3s (4)             2.3s (4)  -1.1%
     sqlalchemy_imperative2.py             2.9s (4)             2.8s (4)  -0.7%
                       geomean                 2.6s                 2.6s  -0.9%
```